refactor: refactor proposal streamer
